fixed mismatched type in parameters

+++ include/local_planner/custom_planner.hpp
@@ -82,7 +82,7 @@ namespace local_planner
         int local_plan_projection_number;
         double vel_reduction_factor;
         double local_plan_rotation_rate_factor;
-        double local_plan_rotation_rate;
+        int local_plan_rotation_rate;
 
 
         typedef std::vector<std::tuple<double, double>> velocities;
